--- process_gharchive.py
@@ -1,123 +1,0 @@
-import glob
-import json
-
-import polars as pl
-from polars.datatypes import (
-    Array,
-    Binary,
-    Boolean,
-    Categorical,
-    DataType,
-    Date,
-    Datetime,
-    Decimal,
-    Duration,
-    Enum,
-    Field,
-    Float32,
-    Float64,
-    Int8,
-    Int16,
-    Int32,
-    Int64,
-    List,
-    Null,
-    Object,
-    String,
-    Struct,
-    Time,
-    UInt8,
-    UInt16,
-    UInt32,
-    UInt64,
-    Unknown,
-    Utf8,
-)
-from polars.schema import Schema
-
-
-# 0. Loading the data
-schema = Schema(
-    [
-        ("id", String),
-        ("type", String),
-        (
-            "actor",
-            Struct(
-                {
-                    "id": Int64,
-                    "login": String,
-                    "display_login": String,
-                    "gravatar_id": String,
-                    "url": String,
-                    "avatar_url": String,
-                }
-            ),
-        ),
-        ("repo", Struct({"id": Int64, "name": String, "url": String})),
-        (
-            "payload",
-            Struct(
-                {
-                    "repository_id": Int64,
-                    "push_id": Int64,
-                    "size": Int64,
-                    "distinct_size": Int64,
-                    "ref": String,
-                    "head": String,
-                    "before": String,
-                    "commits": List(
-                        Struct(
-                            {
-                                "sha": String,
-                                "author": Struct({"email": String, "name": String}),
-                                "message": String,
-                                "distinct": Boolean,
-                                "url": String,
-                            }
-                        )
-                    ),
-                    "action": String,
-                    "number": Int64,
-                    "pull_request": Struct(
-                        {
-                            "state": String,
-                            "user": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "merged": Boolean,
-                        }
-                    ),
-                }
-            ),
-        ),
-        ("public", String),
-        ("created_at", String),
-    ]
-)
-
-df = pl.read_ndjson(
-    "node_input/node.json",
-    schema=schema,
-)
-
-print("Done!!!")
