--- schema.py
@@ -1,1267 +1,0 @@
-import polars as pl
-from polars.datatypes import (
-    Array,
-    Binary,
-    Boolean,
-    Categorical,
-    DataType,
-    Date,
-    Datetime,
-    Decimal,
-    Duration,
-    Enum,
-    Field,
-    Float32,
-    Float64,
-    Int8,
-    Int16,
-    Int32,
-    Int64,
-    List,
-    Null,
-    Object,
-    String,
-    Struct,
-    Time,
-    UInt8,
-    UInt16,
-    UInt32,
-    UInt64,
-    Unknown,
-    Utf8,
-)
-from polars.schema import Schema
-
-schema = Schema(
-    [
-        ("id", String),
-        ("type", String),
-        (
-            "actor",
-            Struct(
-                {
-                    "id": Int64,
-                    "login": String,
-                    "display_login": String,
-                    "gravatar_id": String,
-                    "url": String,
-                    "avatar_url": String,
-                }
-            ),
-        ),
-        ("repo", Struct({"id": Int64, "name": String, "url": String})),
-        (
-            "payload",
-            Struct(
-                {
-                    "repository_id": Int64,
-                    "push_id": Int64,
-                    "size": Int64,
-                    "distinct_size": Int64,
-                    "ref": String,
-                    "head": String,
-                    "before": String,
-                    "commits": List(
-                        Struct(
-                            {
-                                "sha": String,
-                                "author": Struct({"email": String, "name": String}),
-                                "message": String,
-                                "distinct": Boolean,
-                                "url": String,
-                            }
-                        )
-                    ),
-                    "action": String,
-                    "number": Int64,
-                    "pull_request": Struct(
-                        {
-                            "labels": List(
-                                Struct(
-                                    {
-                                        "id": Int64,
-                                        "node_id": String,
-                                        "url": String,
-                                        "name": String,
-                                        "color": String,
-                                        "default": Boolean,
-                                        "description": String,
-                                    }
-                                )
-                            ),
-                            "review_comments_url": String,
-                            "state": String,
-                            "user": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "locked": Boolean,
-                            "merged_at": String,
-                            "diff_url": String,
-                            "assignees": List(
-                                Struct(
-                                    {
-                                        "login": String,
-                                        "id": Int64,
-                                        "node_id": String,
-                                        "avatar_url": String,
-                                        "gravatar_id": String,
-                                        "url": String,
-                                        "html_url": String,
-                                        "followers_url": String,
-                                        "following_url": String,
-                                        "gists_url": String,
-                                        "starred_url": String,
-                                        "subscriptions_url": String,
-                                        "organizations_url": String,
-                                        "repos_url": String,
-                                        "events_url": String,
-                                        "received_events_url": String,
-                                        "type": String,
-                                        "site_admin": Boolean,
-                                    }
-                                )
-                            ),
-                            "requested_teams": List(
-                                Struct(
-                                    {
-                                        "name": String,
-                                        "id": Int64,
-                                        "node_id": String,
-                                        "slug": String,
-                                        "description": String,
-                                        "privacy": String,
-                                        "notification_setting": String,
-                                        "url": String,
-                                        "html_url": String,
-                                        "members_url": String,
-                                        "repositories_url": String,
-                                        "permission": String,
-                                        "parent": Null,
-                                    }
-                                )
-                            ),
-                            "review_comment_url": String,
-                            "auto_merge": Struct(
-                                {
-                                    "enabled_by": Struct(
-                                        {
-                                            "login": String,
-                                            "id": Int64,
-                                            "node_id": String,
-                                            "avatar_url": String,
-                                            "gravatar_id": String,
-                                            "url": String,
-                                            "html_url": String,
-                                            "followers_url": String,
-                                            "following_url": String,
-                                            "gists_url": String,
-                                            "starred_url": String,
-                                            "subscriptions_url": String,
-                                            "organizations_url": String,
-                                            "repos_url": String,
-                                            "events_url": String,
-                                            "received_events_url": String,
-                                            "type": String,
-                                            "site_admin": Boolean,
-                                        }
-                                    ),
-                                    "merge_method": String,
-                                    "commit_title": String,
-                                    "commit_message": String,
-                                }
-                            ),
-                            "statuses_url": String,
-                            "patch_url": String,
-                            "mergeable_state": String,
-                            "comments": Int64,
-                            "_links": Struct(
-                                {
-                                    "self": Struct({"href": String}),
-                                    "html": Struct({"href": String}),
-                                    "issue": Struct({"href": String}),
-                                    "comments": Struct({"href": String}),
-                                    "review_comments": Struct({"href": String}),
-                                    "review_comment": Struct({"href": String}),
-                                    "commits": Struct({"href": String}),
-                                    "statuses": Struct({"href": String}),
-                                }
-                            ),
-                            "additions": Int64,
-                            "issue_url": String,
-                            "url": String,
-                            "number": Int64,
-                            "closed_at": String,
-                            "merge_commit_sha": String,
-                            "milestone": Struct(
-                                {
-                                    "url": String,
-                                    "html_url": String,
-                                    "labels_url": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "number": Int64,
-                                    "title": String,
-                                    "description": String,
-                                    "creator": Struct(
-                                        {
-                                            "login": String,
-                                            "id": Int64,
-                                            "node_id": String,
-                                            "avatar_url": String,
-                                            "gravatar_id": String,
-                                            "url": String,
-                                            "html_url": String,
-                                            "followers_url": String,
-                                            "following_url": String,
-                                            "gists_url": String,
-                                            "starred_url": String,
-                                            "subscriptions_url": String,
-                                            "organizations_url": String,
-                                            "repos_url": String,
-                                            "events_url": String,
-                                            "received_events_url": String,
-                                            "type": String,
-                                            "site_admin": Boolean,
-                                        }
-                                    ),
-                                    "open_issues": Int64,
-                                    "closed_issues": Int64,
-                                    "state": String,
-                                    "created_at": String,
-                                    "updated_at": String,
-                                    "due_on": String,
-                                    "closed_at": Null,
-                                }
-                            ),
-                            "node_id": String,
-                            "comments_url": String,
-                            "mergeable": Boolean,
-                            "body": String,
-                            "requested_reviewers": List(
-                                Struct(
-                                    {
-                                        "login": String,
-                                        "id": Int64,
-                                        "node_id": String,
-                                        "avatar_url": String,
-                                        "gravatar_id": String,
-                                        "url": String,
-                                        "html_url": String,
-                                        "followers_url": String,
-                                        "following_url": String,
-                                        "gists_url": String,
-                                        "starred_url": String,
-                                        "subscriptions_url": String,
-                                        "organizations_url": String,
-                                        "repos_url": String,
-                                        "events_url": String,
-                                        "received_events_url": String,
-                                        "type": String,
-                                        "site_admin": Boolean,
-                                    }
-                                )
-                            ),
-                            "rebaseable": Boolean,
-                            "updated_at": String,
-                            "title": String,
-                            "base": Struct(
-                                {
-                                    "label": String,
-                                    "ref": String,
-                                    "sha": String,
-                                    "user": Struct(
-                                        {
-                                            "login": String,
-                                            "id": Int64,
-                                            "node_id": String,
-                                            "avatar_url": String,
-                                            "gravatar_id": String,
-                                            "url": String,
-                                            "html_url": String,
-                                            "followers_url": String,
-                                            "following_url": String,
-                                            "gists_url": String,
-                                            "starred_url": String,
-                                            "subscriptions_url": String,
-                                            "organizations_url": String,
-                                            "repos_url": String,
-                                            "events_url": String,
-                                            "received_events_url": String,
-                                            "type": String,
-                                            "site_admin": Boolean,
-                                        }
-                                    ),
-                                    "repo": Struct(
-                                        {
-                                            "forks_count": Int64,
-                                            "issue_comment_url": String,
-                                            "allow_forking": Boolean,
-                                            "languages_url": String,
-                                            "visibility": String,
-                                            "disabled": Boolean,
-                                            "statuses_url": String,
-                                            "language": String,
-                                            "labels_url": String,
-                                            "has_projects": Boolean,
-                                            "open_issues": Int64,
-                                            "url": String,
-                                            "downloads_url": String,
-                                            "tags_url": String,
-                                            "trees_url": String,
-                                            "watchers_count": Int64,
-                                            "comments_url": String,
-                                            "has_issues": Boolean,
-                                            "has_pages": Boolean,
-                                            "subscribers_url": String,
-                                            "has_downloads": Boolean,
-                                            "watchers": Int64,
-                                            "updated_at": String,
-                                            "commits_url": String,
-                                            "pushed_at": String,
-                                            "blobs_url": String,
-                                            "issues_url": String,
-                                            "description": String,
-                                            "license": Struct(
-                                                {
-                                                    "key": String,
-                                                    "name": String,
-                                                    "spdx_id": String,
-                                                    "url": String,
-                                                    "node_id": String,
-                                                }
-                                            ),
-                                            "size": Int64,
-                                            "teams_url": String,
-                                            "compare_url": String,
-                                            "archived": Boolean,
-                                            "stargazers_count": Int64,
-                                            "is_template": Boolean,
-                                            "git_url": String,
-                                            "has_discussions": Boolean,
-                                            "default_branch": String,
-                                            "git_refs_url": String,
-                                            "contents_url": String,
-                                            "web_commit_signoff_required": Boolean,
-                                            "issue_events_url": String,
-                                            "hooks_url": String,
-                                            "fork": Boolean,
-                                            "events_url": String,
-                                            "forks": Int64,
-                                            "assignees_url": String,
-                                            "svn_url": String,
-                                            "keys_url": String,
-                                            "mirror_url": Null,
-                                            "node_id": String,
-                                            "git_commits_url": String,
-                                            "pulls_url": String,
-                                            "merges_url": String,
-                                            "contributors_url": String,
-                                            "private": Boolean,
-                                            "open_issues_count": Int64,
-                                            "topics": List(String),
-                                            "stargazers_url": String,
-                                            "has_wiki": Boolean,
-                                            "archive_url": String,
-                                            "branches_url": String,
-                                            "forks_url": String,
-                                            "collaborators_url": String,
-                                            "full_name": String,
-                                            "id": Int64,
-                                            "name": String,
-                                            "html_url": String,
-                                            "subscription_url": String,
-                                            "owner": Struct(
-                                                {
-                                                    "login": String,
-                                                    "id": Int64,
-                                                    "node_id": String,
-                                                    "avatar_url": String,
-                                                    "gravatar_id": String,
-                                                    "url": String,
-                                                    "html_url": String,
-                                                    "followers_url": String,
-                                                    "following_url": String,
-                                                    "gists_url": String,
-                                                    "starred_url": String,
-                                                    "subscriptions_url": String,
-                                                    "organizations_url": String,
-                                                    "repos_url": String,
-                                                    "events_url": String,
-                                                    "received_events_url": String,
-                                                    "type": String,
-                                                    "site_admin": Boolean,
-                                                }
-                                            ),
-                                            "git_tags_url": String,
-                                            "milestones_url": String,
-                                            "releases_url": String,
-                                            "created_at": String,
-                                            "ssh_url": String,
-                                            "clone_url": String,
-                                            "deployments_url": String,
-                                            "notifications_url": String,
-                                            "homepage": String,
-                                        }
-                                    ),
-                                }
-                            ),
-                            "commits_url": String,
-                            "author_association": String,
-                            "head": Struct(
-                                {
-                                    "label": String,
-                                    "ref": String,
-                                    "sha": String,
-                                    "user": Struct(
-                                        {
-                                            "login": String,
-                                            "id": Int64,
-                                            "node_id": String,
-                                            "avatar_url": String,
-                                            "gravatar_id": String,
-                                            "url": String,
-                                            "html_url": String,
-                                            "followers_url": String,
-                                            "following_url": String,
-                                            "gists_url": String,
-                                            "starred_url": String,
-                                            "subscriptions_url": String,
-                                            "organizations_url": String,
-                                            "repos_url": String,
-                                            "events_url": String,
-                                            "received_events_url": String,
-                                            "type": String,
-                                            "site_admin": Boolean,
-                                        }
-                                    ),
-                                    "repo": Struct(
-                                        {
-                                            "forks_count": Int64,
-                                            "issue_comment_url": String,
-                                            "allow_forking": Boolean,
-                                            "languages_url": String,
-                                            "visibility": String,
-                                            "disabled": Boolean,
-                                            "statuses_url": String,
-                                            "language": String,
-                                            "labels_url": String,
-                                            "has_projects": Boolean,
-                                            "open_issues": Int64,
-                                            "url": String,
-                                            "downloads_url": String,
-                                            "tags_url": String,
-                                            "trees_url": String,
-                                            "watchers_count": Int64,
-                                            "comments_url": String,
-                                            "has_issues": Boolean,
-                                            "has_pages": Boolean,
-                                            "subscribers_url": String,
-                                            "has_downloads": Boolean,
-                                            "watchers": Int64,
-                                            "updated_at": String,
-                                            "commits_url": String,
-                                            "pushed_at": String,
-                                            "blobs_url": String,
-                                            "issues_url": String,
-                                            "description": String,
-                                            "license": Struct(
-                                                {
-                                                    "key": String,
-                                                    "name": String,
-                                                    "spdx_id": String,
-                                                    "url": String,
-                                                    "node_id": String,
-                                                }
-                                            ),
-                                            "size": Int64,
-                                            "teams_url": String,
-                                            "compare_url": String,
-                                            "archived": Boolean,
-                                            "stargazers_count": Int64,
-                                            "is_template": Boolean,
-                                            "git_url": String,
-                                            "has_discussions": Boolean,
-                                            "default_branch": String,
-                                            "git_refs_url": String,
-                                            "contents_url": String,
-                                            "web_commit_signoff_required": Boolean,
-                                            "issue_events_url": String,
-                                            "hooks_url": String,
-                                            "fork": Boolean,
-                                            "events_url": String,
-                                            "forks": Int64,
-                                            "assignees_url": String,
-                                            "svn_url": String,
-                                            "keys_url": String,
-                                            "mirror_url": Null,
-                                            "node_id": String,
-                                            "git_commits_url": String,
-                                            "pulls_url": String,
-                                            "merges_url": String,
-                                            "contributors_url": String,
-                                            "private": Boolean,
-                                            "open_issues_count": Int64,
-                                            "topics": List(String),
-                                            "stargazers_url": String,
-                                            "has_wiki": Boolean,
-                                            "archive_url": String,
-                                            "branches_url": String,
-                                            "forks_url": String,
-                                            "collaborators_url": String,
-                                            "full_name": String,
-                                            "id": Int64,
-                                            "name": String,
-                                            "html_url": String,
-                                            "subscription_url": String,
-                                            "owner": Struct(
-                                                {
-                                                    "login": String,
-                                                    "id": Int64,
-                                                    "node_id": String,
-                                                    "avatar_url": String,
-                                                    "gravatar_id": String,
-                                                    "url": String,
-                                                    "html_url": String,
-                                                    "followers_url": String,
-                                                    "following_url": String,
-                                                    "gists_url": String,
-                                                    "starred_url": String,
-                                                    "subscriptions_url": String,
-                                                    "organizations_url": String,
-                                                    "repos_url": String,
-                                                    "events_url": String,
-                                                    "received_events_url": String,
-                                                    "type": String,
-                                                    "site_admin": Boolean,
-                                                }
-                                            ),
-                                            "git_tags_url": String,
-                                            "milestones_url": String,
-                                            "releases_url": String,
-                                            "created_at": String,
-                                            "ssh_url": String,
-                                            "clone_url": String,
-                                            "deployments_url": String,
-                                            "notifications_url": String,
-                                            "homepage": String,
-                                        }
-                                    ),
-                                }
-                            ),
-                            "merged": Boolean,
-                            "maintainer_can_modify": Boolean,
-                            "merged_by": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "commits": Int64,
-                            "deletions": Int64,
-                            "changed_files": Int64,
-                            "html_url": String,
-                            "id": Int64,
-                            "created_at": String,
-                            "review_comments": Int64,
-                            "active_lock_reason": Null,
-                            "draft": Boolean,
-                            "assignee": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                        }
-                    ),
-                    "issue": Struct(
-                        {
-                            "url": String,
-                            "repository_url": String,
-                            "labels_url": String,
-                            "comments_url": String,
-                            "events_url": String,
-                            "html_url": String,
-                            "id": Int64,
-                            "node_id": String,
-                            "number": Int64,
-                            "title": String,
-                            "user": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "labels": List(
-                                Struct(
-                                    {
-                                        "id": Int64,
-                                        "node_id": String,
-                                        "url": String,
-                                        "name": String,
-                                        "color": String,
-                                        "default": Boolean,
-                                        "description": String,
-                                    }
-                                )
-                            ),
-                            "state": String,
-                            "locked": Boolean,
-                            "assignee": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "assignees": List(
-                                Struct(
-                                    {
-                                        "login": String,
-                                        "id": Int64,
-                                        "node_id": String,
-                                        "avatar_url": String,
-                                        "gravatar_id": String,
-                                        "url": String,
-                                        "html_url": String,
-                                        "followers_url": String,
-                                        "following_url": String,
-                                        "gists_url": String,
-                                        "starred_url": String,
-                                        "subscriptions_url": String,
-                                        "organizations_url": String,
-                                        "repos_url": String,
-                                        "events_url": String,
-                                        "received_events_url": String,
-                                        "type": String,
-                                        "site_admin": Boolean,
-                                    }
-                                )
-                            ),
-                            "milestone": Struct(
-                                {
-                                    "url": String,
-                                    "html_url": String,
-                                    "labels_url": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "number": Int64,
-                                    "title": String,
-                                    "description": String,
-                                    "creator": Struct(
-                                        {
-                                            "login": String,
-                                            "id": Int64,
-                                            "node_id": String,
-                                            "avatar_url": String,
-                                            "gravatar_id": String,
-                                            "url": String,
-                                            "html_url": String,
-                                            "followers_url": String,
-                                            "following_url": String,
-                                            "gists_url": String,
-                                            "starred_url": String,
-                                            "subscriptions_url": String,
-                                            "organizations_url": String,
-                                            "repos_url": String,
-                                            "events_url": String,
-                                            "received_events_url": String,
-                                            "type": String,
-                                            "site_admin": Boolean,
-                                        }
-                                    ),
-                                    "open_issues": Int64,
-                                    "closed_issues": Int64,
-                                    "state": String,
-                                    "created_at": String,
-                                    "updated_at": String,
-                                    "due_on": String,
-                                    "closed_at": String,
-                                }
-                            ),
-                            "comments": Int64,
-                            "created_at": String,
-                            "updated_at": String,
-                            "closed_at": String,
-                            "author_association": String,
-                            "active_lock_reason": String,
-                            "draft": Boolean,
-                            "pull_request": Struct(
-                                {
-                                    "url": String,
-                                    "html_url": String,
-                                    "diff_url": String,
-                                    "patch_url": String,
-                                    "merged_at": String,
-                                }
-                            ),
-                            "body": String,
-                            "reactions": Struct(
-                                {
-                                    "url": String,
-                                    "total_count": Int64,
-                                    "+1": Int64,
-                                    "-1": Int64,
-                                    "laugh": Int64,
-                                    "hooray": Int64,
-                                    "confused": Int64,
-                                    "heart": Int64,
-                                    "rocket": Int64,
-                                    "eyes": Int64,
-                                }
-                            ),
-                            "timeline_url": String,
-                            "performed_via_github_app": Null,
-                            "state_reason": String,
-                        }
-                    ),
-                    "ref_type": String,
-                    "pusher_type": String,
-                    "master_branch": String,
-                    "description": String,
-                    "comment": Struct(
-                        {
-                            "url": String,
-                            "pull_request_review_id": Int64,
-                            "id": Int64,
-                            "node_id": String,
-                            "diff_hunk": String,
-                            "path": String,
-                            "commit_id": String,
-                            "original_commit_id": String,
-                            "user": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "body": String,
-                            "created_at": String,
-                            "updated_at": String,
-                            "html_url": String,
-                            "pull_request_url": String,
-                            "author_association": String,
-                            "_links": Struct(
-                                {
-                                    "self": Struct({"href": String}),
-                                    "html": Struct({"href": String}),
-                                    "pull_request": Struct({"href": String}),
-                                }
-                            ),
-                            "reactions": Struct(
-                                {
-                                    "url": String,
-                                    "total_count": Int64,
-                                    "+1": Int64,
-                                    "-1": Int64,
-                                    "laugh": Int64,
-                                    "hooray": Int64,
-                                    "confused": Int64,
-                                    "heart": Int64,
-                                    "rocket": Int64,
-                                    "eyes": Int64,
-                                }
-                            ),
-                            "start_line": Int64,
-                            "original_start_line": Int64,
-                            "start_side": String,
-                            "line": Int64,
-                            "original_line": Int64,
-                            "side": String,
-                            "original_position": Int64,
-                            "position": Int64,
-                            "subject_type": String,
-                            "issue_url": String,
-                            "performed_via_github_app": Struct(
-                                {
-                                    "id": Int64,
-                                    "slug": String,
-                                    "node_id": String,
-                                    "owner": Struct(
-                                        {
-                                            "login": String,
-                                            "id": Int64,
-                                            "node_id": String,
-                                            "avatar_url": String,
-                                            "gravatar_id": String,
-                                            "url": String,
-                                            "html_url": String,
-                                            "followers_url": String,
-                                            "following_url": String,
-                                            "gists_url": String,
-                                            "starred_url": String,
-                                            "subscriptions_url": String,
-                                            "organizations_url": String,
-                                            "repos_url": String,
-                                            "events_url": String,
-                                            "received_events_url": String,
-                                            "type": String,
-                                            "site_admin": Boolean,
-                                        }
-                                    ),
-                                    "name": String,
-                                    "description": String,
-                                    "external_url": String,
-                                    "html_url": String,
-                                    "created_at": String,
-                                    "updated_at": String,
-                                    "permissions": Struct(
-                                        {
-                                            "organization_hooks": String,
-                                            "organization_secrets": String,
-                                            "gpg_keys": String,
-                                            "watching": String,
-                                            "issues": String,
-                                            "secrets": String,
-                                            "interaction_limits": String,
-                                            "deployments": String,
-                                            "administration": String,
-                                            "organization_projects": String,
-                                            "packages": String,
-                                            "pull_requests": String,
-                                            "workflows": String,
-                                            "organization_dependabot_secrets": String,
-                                            "followers": String,
-                                            "members": String,
-                                            "starring": String,
-                                            "organization_self_hosted_runners": String,
-                                            "keys": String,
-                                            "repository_hooks": String,
-                                            "dependabot_secrets": String,
-                                            "checks": String,
-                                            "environments": String,
-                                            "pages": String,
-                                            "organization_plan": String,
-                                            "repository_projects": String,
-                                            "organization_events": String,
-                                            "blocking": String,
-                                            "discussions": String,
-                                            "actions": String,
-                                            "organization_administration": String,
-                                            "plan": String,
-                                            "secret_scanning_alerts": String,
-                                            "vulnerability_alerts": String,
-                                            "security_events": String,
-                                            "contents": String,
-                                            "team_discussions": String,
-                                            "emails": String,
-                                            "statuses": String,
-                                            "organization_user_blocking": String,
-                                            "metadata": String,
-                                            "merge_queues": String,
-                                            "codespaces": String,
-                                            "actions_variables": String,
-                                            "organization_custom_roles": String,
-                                            "codespaces_metadata": String,
-                                            "single_file": String,
-                                            "gists": String,
-                                            "codespaces_user_secrets": String,
-                                        }
-                                    ),
-                                    "events": List(String),
-                                }
-                            ),
-                            "in_reply_to_id": Int64,
-                        }
-                    ),
-                    "release": Struct(
-                        {
-                            "url": String,
-                            "assets_url": String,
-                            "upload_url": String,
-                            "html_url": String,
-                            "id": Int64,
-                            "author": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "node_id": String,
-                            "tag_name": String,
-                            "target_commitish": String,
-                            "name": String,
-                            "draft": Boolean,
-                            "prerelease": Boolean,
-                            "created_at": String,
-                            "published_at": String,
-                            "assets": List(
-                                Struct(
-                                    {
-                                        "url": String,
-                                        "id": Int64,
-                                        "node_id": String,
-                                        "name": String,
-                                        "label": String,
-                                        "uploader": Struct(
-                                            {
-                                                "login": String,
-                                                "id": Int64,
-                                                "node_id": String,
-                                                "avatar_url": String,
-                                                "gravatar_id": String,
-                                                "url": String,
-                                                "html_url": String,
-                                                "followers_url": String,
-                                                "following_url": String,
-                                                "gists_url": String,
-                                                "starred_url": String,
-                                                "subscriptions_url": String,
-                                                "organizations_url": String,
-                                                "repos_url": String,
-                                                "events_url": String,
-                                                "received_events_url": String,
-                                                "type": String,
-                                                "site_admin": Boolean,
-                                            }
-                                        ),
-                                        "content_type": String,
-                                        "state": String,
-                                        "size": Int64,
-                                        "download_count": Int64,
-                                        "created_at": String,
-                                        "updated_at": String,
-                                        "browser_download_url": String,
-                                    }
-                                )
-                            ),
-                            "tarball_url": String,
-                            "zipball_url": String,
-                            "body": String,
-                            "mentions_count": Int64,
-                            "mentions": List(
-                                Struct(
-                                    {
-                                        "avatar_url": String,
-                                        "login": String,
-                                        "profile_name": String,
-                                        "profile_url": String,
-                                        "avatar_user_actor": Boolean,
-                                    }
-                                )
-                            ),
-                            "short_description_html": String,
-                            "is_short_description_html_truncated": Boolean,
-                            "discussion_url": String,
-                        }
-                    ),
-                    "forkee": Struct(
-                        {
-                            "forks_count": Int64,
-                            "issue_comment_url": String,
-                            "allow_forking": Boolean,
-                            "languages_url": String,
-                            "visibility": String,
-                            "disabled": Boolean,
-                            "statuses_url": String,
-                            "language": Null,
-                            "labels_url": String,
-                            "has_projects": Boolean,
-                            "open_issues": Int64,
-                            "url": String,
-                            "downloads_url": String,
-                            "tags_url": String,
-                            "trees_url": String,
-                            "watchers_count": Int64,
-                            "comments_url": String,
-                            "has_issues": Boolean,
-                            "has_pages": Boolean,
-                            "subscribers_url": String,
-                            "has_downloads": Boolean,
-                            "watchers": Int64,
-                            "updated_at": String,
-                            "commits_url": String,
-                            "pushed_at": String,
-                            "blobs_url": String,
-                            "issues_url": String,
-                            "description": String,
-                            "license": Struct(
-                                {
-                                    "key": String,
-                                    "name": String,
-                                    "spdx_id": String,
-                                    "url": String,
-                                    "node_id": String,
-                                }
-                            ),
-                            "size": Int64,
-                            "teams_url": String,
-                            "compare_url": String,
-                            "archived": Boolean,
-                            "stargazers_count": Int64,
-                            "is_template": Boolean,
-                            "git_url": String,
-                            "has_discussions": Boolean,
-                            "default_branch": String,
-                            "public": Boolean,
-                            "git_refs_url": String,
-                            "contents_url": String,
-                            "web_commit_signoff_required": Boolean,
-                            "issue_events_url": String,
-                            "hooks_url": String,
-                            "fork": Boolean,
-                            "events_url": String,
-                            "forks": Int64,
-                            "assignees_url": String,
-                            "svn_url": String,
-                            "keys_url": String,
-                            "mirror_url": Null,
-                            "node_id": String,
-                            "git_commits_url": String,
-                            "pulls_url": String,
-                            "merges_url": String,
-                            "contributors_url": String,
-                            "private": Boolean,
-                            "open_issues_count": Int64,
-                            "topics": List(Null),
-                            "stargazers_url": String,
-                            "has_wiki": Boolean,
-                            "archive_url": String,
-                            "branches_url": String,
-                            "forks_url": String,
-                            "collaborators_url": String,
-                            "full_name": String,
-                            "id": Int64,
-                            "name": String,
-                            "html_url": String,
-                            "subscription_url": String,
-                            "owner": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "git_tags_url": String,
-                            "milestones_url": String,
-                            "releases_url": String,
-                            "created_at": String,
-                            "ssh_url": String,
-                            "clone_url": String,
-                            "deployments_url": String,
-                            "notifications_url": String,
-                            "homepage": String,
-                        }
-                    ),
-                    "review": Struct(
-                        {
-                            "id": Int64,
-                            "node_id": String,
-                            "user": Struct(
-                                {
-                                    "login": String,
-                                    "id": Int64,
-                                    "node_id": String,
-                                    "avatar_url": String,
-                                    "gravatar_id": String,
-                                    "url": String,
-                                    "html_url": String,
-                                    "followers_url": String,
-                                    "following_url": String,
-                                    "gists_url": String,
-                                    "starred_url": String,
-                                    "subscriptions_url": String,
-                                    "organizations_url": String,
-                                    "repos_url": String,
-                                    "events_url": String,
-                                    "received_events_url": String,
-                                    "type": String,
-                                    "site_admin": Boolean,
-                                }
-                            ),
-                            "body": String,
-                            "commit_id": String,
-                            "submitted_at": String,
-                            "state": String,
-                            "html_url": String,
-                            "pull_request_url": String,
-                            "author_association": String,
-                            "_links": Struct(
-                                {
-                                    "html": Struct({"href": String}),
-                                    "pull_request": Struct({"href": String}),
-                                }
-                            ),
-                        }
-                    ),
-                    "member": Struct(
-                        {
-                            "login": String,
-                            "id": Int64,
-                            "node_id": String,
-                            "avatar_url": String,
-                            "gravatar_id": String,
-                            "url": String,
-                            "html_url": String,
-                            "followers_url": String,
-                            "following_url": String,
-                            "gists_url": String,
-                            "starred_url": String,
-                            "subscriptions_url": String,
-                            "organizations_url": String,
-                            "repos_url": String,
-                            "events_url": String,
-                            "received_events_url": String,
-                            "type": String,
-                            "site_admin": Boolean,
-                        }
-                    ),
-                    "pages": List(
-                        Struct(
-                            {
-                                "page_name": String,
-                                "title": String,
-                                "summary": Null,
-                                "action": String,
-                                "sha": String,
-                                "html_url": String,
-                            }
-                        )
-                    ),
-                }
-            ),
-        ),
-        ("public", String),
-        ("created_at", String),
-        (
-            "org",
-            Struct(
-                {
-                    "id": Int64,
-                    "login": String,
-                    "gravatar_id": String,
-                    "url": String,
-                    "avatar_url": String,
-                }
-            ),
-        ),
-    ]
-)
